requirements.txt
python-socketio==3.1.2
aiohttp==3.7.4.post0
python-engineio==3.13.2
tornado
watchgod==0.7
pymongo==3.11.4
python-dotenv==0.17.1
dynaconf==3.1.4
motor==2.4.0





